+++ RELEASE-NOTES.md
@@ -5,6 +5,7 @@ Release Notes
 -----
 - changed package names to match Maven group id
 - using copied data set to produce JSON data to avoid ConcurrentModificationException (happens while saving data to file and changing store data simultaneously)
+- added slf4j logging api
 
 0.1.0
 -----